--- app/src/main/java/com/chrisvasqm/cuadramo/data/source/user/UserService.kt
@@ -1,12 +1,0 @@
-package com.chrisvasqm.cuadramo.data.source.user
-
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.FirebaseUser
-
-class UserService {
-
-    fun getCurrentUser(): FirebaseUser? {
-        return FirebaseAuth.getInstance().currentUser
-    }
-
-}
