+++ app/src/main/java/com/chrisvasqm/cuadramo/ui/PreviewBottomSheetDialogFragment.kt
@@ -6,11 +6,10 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.FragmentManager
-import com.chrisvasqm.cuadramo.data.models.Cuadre
+import com.chrisvasqm.cuadramo.model.Cuadre
 import com.chrisvasqm.cuadramo.databinding.BottomDialogCuadrePreviewBinding
 import com.chrisvasqm.cuadramo.extensions.toString
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
-import com.google.firebase.auth.FirebaseAuth
 
 class PreviewBottomSheetDialogFragment : BottomSheetDialogFragment() {
